Footer 가운데 정렬, 파일 정리
