+++ creotest/src/App.css
@@ -18,7 +18,7 @@
 .wrap {
   min-height: 100%;
   position: relative;
-  padding-bottom: 400px;
+  padding-bottom: 100px;
   /* background-color: pink; */
   /* border: 1px solid red; */
 }
@@ -44,7 +44,6 @@
   margin: 15px 0;
   width: 90px;
   object-fit: contain;
-  /* border: 1px solid palevioletred; */
 }
 
 a {  
@@ -63,35 +62,35 @@ a {
   min-height: calc(100% - 170px);
   font-size: 10px;
   color: black;
-  /* border: 1px solid #ff9900; */
+  border: 1px solid #ff9900;
 }
 
 /* Footer */
 .Footer {
   /* display: flex; */
-  position: absolute;
+  position: relative;
   bottom: 0;
   left: 0;
   width: 100%;
-  /* right: 0; */
-  /* padding: 5px; */
   background-color: black;
   color: white;
+  text-align: center;
 }
 
 .FooterUp {
-  position: relative;
+  display: inline-block;
   left: 100px;
-  width: 60%;
+  width: 85%;
   padding: 5px;
   border: 1px solid blue;
+  text-align: left;
 }
 
 .FooterDown {
   /* height: 70px;
   width: 100%; */
-  /* position: inherit;
-  bottom: 0; */
+  position: inherit;
+  bottom: 0;
   display: flex;
   background-color: rgb(31, 31, 31);
   width: 100%;
