+++ creotest/src/App.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { Provider } from './context';
 import './App.css';
 import { Link, Route, Switch } from 'react-router-dom';
-import { Home, Creo, Team, Works, Careers, Contact2 } from './components';
-import Navbar2 from './components/Nav/Navbar2';
+import { Home, Creo, Team, Works, Careers, Contact } from './components';
+import Navbar from './components/Nav/Navbar';
 import Footer from './components/Footer';
 
 const App = () => {
@@ -12,21 +12,21 @@ const App = () => {
     <Provider>
     <div className="creo">
       <div className="body">
+        <Navbar/>
       <div className="wrap">
-        <Navbar2/>
         <div className="Main">
           <Switch>
             <Route exact path="/" component={Home} />
-            <Route path="/components/Creo" component={Creo} />
-            <Route path="/components/Team" component={Team} />
-            <Route path="/components/Works" component={Works} />
-            <Route path="/components/Careers" component={Careers} />
-            <Route path="/components/Contact2" component={Contact2} />
+            <Route path="/Creo" component={Creo} />
+            <Route path="/Team" component={Team} />
+            <Route path="/Works" component={Works} />
+            <Route path="/Careers" component={Careers} />
+            <Route path="/Contact" component={Contact} />
             <Route path="/"> Not Found </Route>
           </Switch>
         </div>
-      <Footer/>
       </div>
+      <Footer/>
     </div>
   </div>
   </Provider>
