+++ creotest/src/components/Contact.js
@@ -1,23 +1,19 @@
-import React, { Component } from 'react'
+import React, { useState } from 'react';
 import Title from './Title';
+import './Components.scss';
 
-export default class Contact extends Component {
+function Contact2 () {
 
-    state = { name: '', phone: '', email: '', organization: '', message: '' };
+    const [ name, setName ] = useState("");
+    const [ phone, setPhone ] = useState("");
+    const [ email, setEmail ] = useState("");
+    const [ organization, setOrganization ] = useState("");
+    const [ message, setMessage ] = useState("");
 
-    handleChange = (e) => {
-        this.setState({ name: e.target.value, phone: e.target.value, email: e.target.value, organization: e.target.value, message: e.target.value });
+    const onClick = () => {
+        alert( "name: " + name + "/ phone: " + phone + "/ email: " + email + "/ organization: " + organization + "/ message: " + message )
     }
 
-    handleClick = () => {
-        const { name, phone, email, organization, message } = this.state;
-            alert(name + ' ' + phone + ' ' + email + ' ' + organization + ' ' + message);
-            this.setState({ name: '', phone: '', email: '', organizatio: '', message: '' })
-    }
-
-    render() {
-        const { name, phone, email, organization, message } = this.state;
-        const { handleChange, handleClick } = this;
         return (
             <div className = "contact">
                 <Title name = "CONTACT"/>
@@ -26,52 +22,26 @@ export default class Contact extends Component {
                     <p>7, Yeonmujang 5ga-gil,</p> 
                     <p>Seongdong-gu, Seoul, Republic of Korea</p>
                 <h4>HOW CAN WE HELP YOU?</h4>
-
-                <span className="form">
-                    <div className="li">
-                        <span className="li1">Your Name </span>
-                        <input className="box" type="text" alter="name"
-                            value={name} onChange={handleChange}/>
-                    </div>
-                    <div className="li">
-                        <span className="li1">Phone </span>
-                        <input className="box" type="text" alter="Phone"
-                        value={phone} onChange={handleChange}/>
-                    </div>
-                    <div className="li">
-                        <span className="li1">E-mail</span>
-                       <input className="box" type="text" alter="email"
-                        value={email} onChange={handleChange}/>
-                    </div>
-                    <div className="li">
-                        <span className="li1">Organization</span>
-                       <input className="box" type="text" alter="Organization"
-                        value={organization} onChange={handleChange}/>
-                    </div>
-                    <div className="li">
-                        <span className="li5">Message </span>
-                       <textarea className="text" type="box" alter="Message"
-                        value={message} onChange={handleChange}/>
-                    </div>
-                </span>
-                {/* <p>Your Name 
-                    <input className="box" type="text" alter="name"
-                        value={name} onChange={handleChange}/></p>
-                <p>Phone 
-                    <input className="box" type="text" alter="Phone"
-                        value={phone} onChange={handleChange}/></p>
-                <p>E-mail 
-                    <input className="box" type="text" alter="email"
-                        value={email} onChange={handleChange}/></p>
-                <p>Organization 
-                    <input className="box" type="text" alter="Organization"
-                        value={organization} onChange={handleChange}/></p>
-                <p>Message 
-                    <textarea className="text" type="box" alter="Message"
-                        value={message} onChange={handleChange}/></p> */}
-                <button type="submit" onClick={handleClick}>Submit</button>
-                <h2>배열 정리, handelChange 분리</h2>
+                <div className="form">
+                    name:
+                    <input className="box" type="text" value={name} 
+                    onChange={({ target: {value}}) => setName(value)}/>
+                    phone: 
+                    <input className="box" type="text" value={phone} 
+                    onChange={({ target: {value}}) => setPhone(value)}/>
+                    email:
+                    <input className="box" type="text" value={email} 
+                    onChange={({ target: {value}}) => setEmail(value)}/>
+                    organization:
+                    <input className="box" type="text" value={organization} 
+                    onChange={({ target: {value}}) => setOrganization(value)}/>
+                    message:
+                    <input className="message" type="text" value={message} 
+                    onChange={({ target: {value}}) => setMessage(value)}/>
+                    <button className="submit" onClick={onClick}>submit</button>
+                </div>
             </div>
         )
     }
-}
+
+export default Contact2;
