+++ creotest/src/components/index.js
@@ -3,4 +3,4 @@ export { default as Creo } from './Creo';
 export { default as Team } from './Team';
 export { default as Works } from './Works';
 export { default as Careers } from './Careers';
-export { default as Contact2 } from './Contact2';
+export { default as Contact } from './Contact';
